Big 5 Personality Factors

Big 5 Personality Factor data collected online

big5

Format

A data frame with 19719 rows and 57 variables:

race

factor. 13 levels.

age

Age in years.

gender

factor. Coded as "Male", "Female", "Other".

hand

factor. What had do you write with? Coded as "Left", "Right", "Both".

country

character. Techinal location as ISO country code.

E1

I am the life of the party.

E2

I don't talk a lot.

E3

I feel comfortable around people.

E4

I keep in the background.

E5

I start conversations.

E6

I have little to say.

E7

I talk to a lot of different people at parties.

E8

I don't like to draw attention to myself.

E9

I don't mind being the center of attention.

E10

I am quiet around strangers.

N1

I get stressed out easily.

N2

I am relaxed most of the time.

N3

I worry about things.

N4

I seldom feel blue.

N5

I am easily disturbed.

N6

I get upset easily.

N7

I change my mood a lot.

N8

I have frequent mood swings.

N9

I get irritated easily.

N10

I often feel blue.

A1

I feel little concern for others.

A2

I am interested in people.

A3

I insult people.

A4

I sympathize with others' feelings.

A5

I am not interested in other people's problems.

A6

I have a soft heart.

A7

I am not really interested in others.

A8

I take time out for others.

A9

I feel others' emotions.

A10

I make people feel at ease.

C1

I am always prepared.

C2

I leave my belongings around.

C3

I pay attention to details.

C4

I make a mess of things.

C5

I get chores done right away.

C6

I often forget to put things back in their proper place.

C7

I like order.

C8

I shirk my duties.

C9

I follow a schedule.

C10

I am exacting in my work.

O1

I have a rich vocabulary.

O2

I have difficulty understanding abstract ideas.

O3

I have a vivid imagination.

O4

I am not interested in abstract ideas.

O5

I have excellent ideas.

O6

I do not have a good imagination.

O7

I am quick to understand things.

O8

I use difficult words.

O9

I spend time reflecting on things.

O10

I am full of ideas.

Source

Dataset obtained from the Open-Source Psychometrics Project.

Details

This data was collected (c. 2012) through on interactive online personality test. Participants were informed that their responses would be recorded and used for research at the begining of the test and asked to confirm their consent at the end of the test.

The test consists of fifty items that you rate on how true they are about you on a five point scale where 1=Disagree, 3=Neutral and 5=Agree.

The five scales are Openness to experience, Conscientiousness, Extraversion, Agreeableness, and Neuroticism.

References

Goldberg, Lewis R. "The development of markers for the Big-Five factor structure." Psychological assessment 4.1 (1992): 26.

Examples

summary(big5)
#>                                                      race            age       
#>  Caucasian (European)                                  :10537   Min.   :18.00  
#>  Other                                                 : 2553   1st Qu.:20.00  
#>  South East Asian (Chinese, Thai, Malay, Filipino, etc): 1861   Median :25.00  
#>  Caucasian (Indian)                                    : 1518   Mean   :28.92  
#>  Mixed Race                                            : 1434   3rd Qu.:34.00  
#>  (Other)                                               : 1663   Max.   :92.00  
#>  NA's                                                  :  153   NA's   :4060   
#>     gender         hand         country                E1              E2      
#>  Male  : 7608   Right:17424   Length:19719       Min.   :1.000   Min.   :1.00  
#>  Female:11985   Left : 1724   Class :character   1st Qu.:2.000   1st Qu.:2.00  
#>  Other :  102   Both :  471   Mode  :character   Median :3.000   Median :3.00  
#>  NA's  :   24   NA's :  100                      Mean   :2.629   Mean   :2.76  
#>                                                  3rd Qu.:4.000   3rd Qu.:4.00  
#>                                                  Max.   :5.000   Max.   :5.00  
#>                                                  NA's   :1       NA's   :1     
#>        E3              E4              E5              E6       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:3.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:1.000  
#>  Median :4.000   Median :3.000   Median :4.000   Median :2.000  
#>  Mean   :3.417   Mean   :3.152   Mean   :3.432   Mean   :2.453  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:3.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        E7              E8              E9             E10       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:3.000  
#>  Median :3.000   Median :3.000   Median :3.000   Median :4.000  
#>  Mean   :2.867   Mean   :3.377   Mean   :3.094   Mean   :3.586  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:5.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        N1              N2              N3              N4       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:3.000   1st Qu.:2.000  
#>  Median :3.000   Median :3.000   Median :4.000   Median :3.000  
#>  Mean   :3.262   Mean   :3.235   Mean   :3.843   Mean   :2.756  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        N5              N6              N7              N8       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :3.000   Median :3.000   Median :3.000   Median :3.000  
#>  Mean   :2.952   Mean   :2.981   Mean   :3.152   Mean   :2.803  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        N9             N10              A1              A2       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:1.000   1st Qu.:3.000  
#>  Median :3.000   Median :3.000   Median :2.000   Median :4.000  
#>  Mean   :3.135   Mean   :2.834   Mean   :2.312   Mean   :3.927  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:3.000   3rd Qu.:5.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        A3              A4             A5              A6              A7       
#>  Min.   :1.000   Min.   :1.00   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:1.000   1st Qu.:4.00   1st Qu.:1.000   1st Qu.:3.000   1st Qu.:1.000  
#>  Median :2.000   Median :4.00   Median :2.000   Median :4.000   Median :2.000  
#>  Mean   :2.163   Mean   :4.03   Mean   :2.166   Mean   :3.896   Mean   :2.161  
#>  3rd Qu.:3.000   3rd Qu.:5.00   3rd Qu.:3.000   3rd Qu.:5.000   3rd Qu.:3.000  
#>  Max.   :5.000   Max.   :5.00   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1      NA's   :1       NA's   :1       NA's   :1      
#>        A8              A9             A10              C1       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:3.000   1st Qu.:3.000   1st Qu.:3.000   1st Qu.:3.000  
#>  Median :4.000   Median :4.000   Median :4.000   Median :3.000  
#>  Mean   :3.767   Mean   :3.945   Mean   :3.683   Mean   :3.319  
#>  3rd Qu.:5.000   3rd Qu.:5.000   3rd Qu.:5.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        C2              C3              C4              C5       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:3.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :3.000   Median :4.000   Median :3.000   Median :3.000  
#>  Mean   :2.979   Mean   :3.983   Mean   :2.654   Mean   :2.701  
#>  3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:4.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        C6              C7              C8              C9       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:3.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :3.000   Median :4.000   Median :2.000   Median :3.000  
#>  Mean   :2.923   Mean   :3.647   Mean   :2.481   Mean   :3.225  
#>  3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:3.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>       C10              O1              O2             O3              O4      
#>  Min.   :1.000   Min.   :1.000   Min.   :1.00   Min.   :1.000   Min.   :1.00  
#>  1st Qu.:3.000   1st Qu.:3.000   1st Qu.:1.00   1st Qu.:4.000   1st Qu.:1.00  
#>  Median :4.000   Median :4.000   Median :2.00   Median :4.000   Median :2.00  
#>  Mean   :3.637   Mean   :3.693   Mean   :2.15   Mean   :4.126   Mean   :2.08  
#>  3rd Qu.:4.000   3rd Qu.:5.000   3rd Qu.:3.00   3rd Qu.:5.000   3rd Qu.:3.00  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.00   Max.   :5.000   Max.   :5.00  
#>  NA's   :1       NA's   :1       NA's   :1      NA's   :1       NA's   :1     
#>        O5              O6              O7              O8       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:3.000   1st Qu.:1.000   1st Qu.:4.000   1st Qu.:2.000  
#>  Median :4.000   Median :1.000   Median :4.000   Median :3.000  
#>  Mean   :3.873   Mean   :1.795   Mean   :4.073   Mean   :3.208  
#>  3rd Qu.:5.000   3rd Qu.:2.000   3rd Qu.:5.000   3rd Qu.:4.000  
#>  Max.   :5.000   Max.   :5.000   Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1       NA's   :1       NA's   :1      
#>        O9             O10       
#>  Min.   :1.000   Min.   :1.000  
#>  1st Qu.:4.000   1st Qu.:3.000  
#>  Median :4.000   Median :4.000  
#>  Mean   :4.134   Mean   :4.005  
#>  3rd Qu.:5.000   3rd Qu.:5.000  
#>  Max.   :5.000   Max.   :5.000  
#>  NA's   :1       NA's   :1